Refactor modules-runtime into multiple files.
